+++ _dev/js/theme/core/index.js
@@ -1,0 +1,1 @@
+import '@js/theme/core/cart/index';
